Refactor: 인증 관련 부분 auth 디렉토리로 이전
