+++ app/auth/dto/responses.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class UserSigninResponse(BaseModel):
+    access_token: str
+    refresh_token: str
